fastapi_ai_translation/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
requests==2.32.5
python-dotenv
prometheus-fastapi-instrumentator

# AI Translation specific
transformers
torch
sentencepiece
huggingface-hub
